+++ build.gradle
@@ -4,6 +4,7 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.6'
 }
 
+
 group = 'com.talearnt'
 version = '0.0.1-SNAPSHOT'
 
@@ -49,6 +50,30 @@ dependencies {
 	annotationProcessor "jakarta.annotation:jakarta.annotation-api"
 	annotationProcessor "jakarta.persistence:jakarta.persistence-api"
 }
+// queryDSL 설정 시작
+def querydslSrcDir = 'src/main/generated'
+sourceSets {
+	main {
+		java {
+			srcDirs += [ querydslSrcDir ]
+		}
+	}
+}
+
+
+compileJava {
+	options.compilerArgs << '-Aquerydsl.generatedAnnotationClass=javax.annotation.Generated'
+}
+
+tasks.withType(JavaCompile) {
+	options.generatedSourceOutputDirectory = file(querydslSrcDir)
+}
+clean {
+	// clean 실행 시 생성된 QClass 삭제
+	delete file(querydslSrcDir)
+}
+// queryDSL 설정 끝
+
 
 tasks.named('test') {
 	useJUnitPlatform()
