--- src/main/java/com/talearnt/join/agree/Agree.java
@@ -1,43 +1,0 @@
-package com.talearnt.join.agree;
-
-import com.talearnt.user.entity.User;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Setter
-@Getter
-@NoArgsConstructor
-@ToString
-public class Agree {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long agreeNo;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_no")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "code_no")
-    private AgreeCode agreeCode;
-
-    @Column(nullable = false)
-    private boolean isAgree;
-
-    @Column(nullable = false)
-    @CreationTimestamp
-    private LocalDateTime agreeDate;
-
-    @Column(nullable = false,updatable = false)
-    @CreationTimestamp
-    private LocalDateTime createdAt;
-
-}
