--- src/main/java/com/talearnt/join/agree/AgreeCode.java
@@ -1,34 +1,0 @@
-package com.talearnt.join.agree;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Setter
-@Getter
-@NoArgsConstructor
-@ToString
-public class AgreeCode {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long codeNo;
-
-    @Column(nullable = false, length = 10)
-    private String code;
-
-    @Column(nullable = false,updatable = false)
-    @Temporal(TemporalType.TIMESTAMP)
-    @CreationTimestamp //insert 쿼리가 발생했을 때 현재 시간값 적용
-    private LocalDateTime createdAt;
-
-    @Column(nullable = false, columnDefinition = "TINYINT(1) DEFAULT 1")
-    private boolean isActive;
-
-}
